browse: simplify GitRepoInfoTask processing
Processing the results of GitRepoInfoTask is currently quite convoluted,
involving posting a custom event type to the RepoTreeView, which then
processes the event by looking up the model and updating it accordingly.

Simplify by just moving the .apply_data() method from RepoTreeView to
GitRepoModel and connecting it to the task directly.
