+++ cola/models/browse.py
@@ -1,9 +1,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 import time
 
-from qtpy import QtCore
 from qtpy import QtGui
-from qtpy import QtWidgets
 from qtpy.QtCore import Qt
 from qtpy.QtCore import Signal
 
@@ -73,7 +71,6 @@ class GitRepoModel(QtGui.QStandardItemModel):
         cfg = context.cfg
         self.turbo = cfg.get('cola.turbo', False)
         self.default_author = cfg.get('user.name', N_('Author'))
-        self._parent = parent
         self._interesting_paths = set()
         self._interesting_files = set()
         self._runtask = qtutils.RunTask(parent=parent)
@@ -245,8 +242,17 @@ class GitRepoModel(QtGui.QStandardItemModel):
         if self.turbo or path not in self.entries:
             return  # entry doesn't currently exist
         context = self.context
-        task = GitRepoInfoTask(context, self._parent, path, self.default_author)
+        task = GitRepoInfoTask(context, path, self.default_author)
+        task.connect(self.apply_data)
         self._runtask.start(task)
+
+    def apply_data(self, data):
+        entry = self.get(data[0])
+        if entry:
+            entry[1].set_status(data[1])
+            entry[2].setText(data[2])
+            entry[3].setText(data[3])
+            entry[4].setText(data[4])
 
 
 def create_column(col, path, is_dir):
@@ -263,11 +269,10 @@ def create_column(col, path, is_dir):
 class GitRepoInfoTask(qtutils.Task):
     """Handles expensive git lookups for a path."""
 
-    def __init__(self, context, parent, path, default_author):
+    def __init__(self, context, path, default_author):
         qtutils.Task.__init__(self)
         self.context = context
         self.path = path
-        self._parent = parent
         self._default_author = default_author
         self._data = {}
 
@@ -354,25 +359,7 @@ class GitRepoInfoTask(qtutils.Task):
             self.data('author'),
             self.data('date'),
         )
-        app = QtWidgets.QApplication.instance()
-        try:
-            app.postEvent(self._parent, GitRepoInfoEvent(data))
-        except RuntimeError:
-            pass  # The app exited before this task finished
-
-
-class GitRepoInfoEvent(QtCore.QEvent):
-    """Transport mechanism for communicating from a GitRepoInfoTask."""
-
-    # Custom event type
-    TYPE = QtCore.QEvent.Type(QtCore.QEvent.registerEventType())
-
-    def __init__(self, data):
-        QtCore.QEvent.__init__(self, self.TYPE)
-        self.data = data
-
-    def type(self):
-        return self.TYPE
+        return data
 
 
 class GitRepoItem(QtGui.QStandardItem):
